env: moved CARAPACE_DESCRIPTION_LENGTH
