+++ internal/mock/mock.go
@@ -1,4 +1,4 @@
-package common
+package mock
 
 import (
 	"fmt"
